requirements.txt
langchain==0.3.7
langchain-groq==0.2.1
langchain-google-genai==2.0.5
langchain-community==0.3.7
langgraph==0.2.50
streamlit==1.40.1
plotly==5.24.1
pandas==2.2.3
sqlalchemy==2.0.35
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.9.2